end up with core logic without exception handling

+++ services/rating-service/src/main/java/com/Harevich/rating_service/dto/request/RatingRequest.java
@@ -22,7 +22,7 @@ public record RatingRequest(
         @Max(5)
         @Min(1)
         @Schema(description = "rating(stars)", example = "5")
-        byte rating,
+        int rating,
 
         @Schema(description = "voting id",example = "PASSENGER")
         VotingPerson whoVotes,
